Refactored functions into their own files
